Move legacy implementation to v1 subfolder, move new implementation to v2 folder
