+++ packages/components/src/dropdown-menu/v1/style.scss
