+++ packages/components/src/dropdown-menu/v2/stories/index.tsx
@@ -19,7 +19,7 @@ import {
 	DropdownMenuRadioGroup,
 	DropdownMenuRadioItem,
 	DropdownSubMenuTrigger,
-} from '../';
+} from '..';
 
 /**
  * WordPress dependencies
