+++ packages/components/src/style.scss
@@ -24,7 +24,7 @@
 @import "./draggable/style.scss";
 @import "./drop-zone/style.scss";
 @import "./dropdown/style.scss";
-@import "./dropdown-menu/style.scss";
+@import "./dropdown-menu/v1/style.scss";
 @import "./duotone-picker/style.scss";
 @import "./duotone-picker/color-list-picker/style.scss";
 @import "./form-toggle/style.scss";
